requirements.txt
asgiref==3.7.2blinker==1.9.0certifi==2024.12.14charset-normalizer==3.4.1click==8.2.1colorama==0.4.6distlib==0.3.7Django==4.2.6filelock==3.12.4Flask==3.1.1flask-cors==6.0.1idna==3.10itsdangerous==2.2.0Jinja2==3.1.6MarkupSafe==3.0.2mysql-connector-python==9.3.0platformdirs==3.11.0PyMySQL==1.1.1requests==2.32.3sqlparse==0.4.4tzdata==2023.3urllib3==2.3.0virtualenv==20.24.5Werkzeug==3.1.3psycopg2-binary==2.9.9gunicorn